Validaciones

+++ resources/js/components/DatosComponent.vue
@@ -3,6 +3,8 @@
                                     <table style="table-layout: fixed; width: 400%;" class="table">
                                     <tr>
                                         <th>
+                                        <input type="hidden" :value="user.id">
+                                        <input type="hidden" :value="usuario.id">
                                         <input v-if="modoEdicion" id="" class="form-control" type="text" v-model="usuario.name">
                                         <p v-else>{{usuario.name}}</p>
                                         </th>
@@ -18,7 +20,7 @@
                                             </div>
                                             <div v-else>
                                                 <button class="btn btn-primary" type="submit" @click="editar">Editar</button>
-                                                <button class="btn btn-danger" type="submit" @click="modalDelUser">Eliminar</button>
+                                                <button class="btn btn-secondary" type="submit" @click="modalDelUser(usuario.id)">Eliminar</button>
                                             </div>
                                         </td>
                                     </tr>
@@ -27,16 +29,16 @@
   <div class="modal-dialog" role="document">
     <div class="modal-content">
       <div class="modal-header">
-        <h5 class="modal-title" id="exampleModalLabel">Modal title</h5>
+        <h5 class="modal-title" id="exampleModalLabel">Eliminar usuario</h5>
         <button type="button" class="close" data-dismiss="modal" aria-label="Close">
           <span aria-hidden="true">&times;</span>
         </button>
       </div>
       <div class="modal-body">
-        ¿Seguro que deseás eliminar?
+        ¿Seguro que deseás eliminar? | {{this.toDeleteUserId}}
       </div>
       <div class="modal-footer">
-        <button type="button" class="btn btn-primary" @click="delUser">Confirm</button>
+        <button type="button" class="btn btn-primary" @click="delUser(this.toDeleteUserId)">Confirmar</button>
         <button type="button" class="btn btn-danger" data-dismiss="modal">Cancelar</button>
       </div>
     </div>
@@ -46,6 +48,7 @@
 </template>
 
 <script>
+
     export default {
         props: ['usuario'],
         data() {
@@ -75,15 +78,29 @@
                 })
             },
             modalDelUser(){
-                this.toDeleteUserId = this.usuario.name;
-                console.log('toDeleteUserId', this.toDeleteUserId);
+                var toDeleteUserId = this.usuario.id;
+                var usuarioID = this.user.id;
+                console.log(this.toDeleteUserId)
+                console.log(usuarioID)
                 $('#exampleModal').modal({});
             },
-            delUser() {
-                axios.delete(`usuarios/${this.usuario.id}`).then((res) => {
+
+            delUser(id) {
+                var usuarioID = this.user.id;
+                idUser = id;
+                if (idUser === '') {
+                    alert("No existe")
+                }
+                console.log('toDeleteUserId', idUser);
+                if(usuarioID === this.usuario.id){
+                    alert("No puedes eliminar al usuario actual")
+                }else{
+                    axios.delete(`usuarios/${this.usuario.id}`).then((res) => {
                     $('#exampleModal').modal('hide');
                     this.$emit('borrar');
-                });
+                })
+                }
+                
             }
         },
     }
